bats

+++ main.py
@@ -154,7 +154,7 @@ def main():
     model = model
     img = Image.open(f"character/{args.character}.png")
     wRatio = img.size[0] / 256
-    img = img.resize((256, img.size[1] / wRatio))
+    img = img.resize((256, int(img.size[1] / wRatio)))
     input_image = preprocessing_image(img.crop((0, 0, 256, 256))).unsqueeze(0)
     extra_image = None
     if img.size[1]>256:
